Remove code duplication.
