+++ Filtration/ViewModels/ItemFilterCommentBlockViewModel.cs
@@ -1,5 +1,4 @@
 ﻿using Filtration.ObjectModel;
-using GalaSoft.MvvmLight.CommandWpf;
 
 namespace Filtration.ViewModels
 {
@@ -11,31 +10,6 @@ namespace Filtration.ViewModels
 
     internal class ItemFilterCommentBlockViewModel : ItemFilterBlockViewModelBase, IItemFilterCommentBlockViewModel
     {
-        private IItemFilterScriptViewModel _parentScriptViewModel;
-
-        public ItemFilterCommentBlockViewModel()
-        {
-            CopyBlockCommand = new RelayCommand(OnCopyBlockCommand);
-            PasteBlockCommand = new RelayCommand(OnPasteBlockCommand);
-            AddBlockCommand = new RelayCommand(OnAddBlockCommand);
-            AddSectionCommand = new RelayCommand(OnAddSectionCommand);
-            DeleteBlockCommand = new RelayCommand(OnDeleteBlockCommand);
-            MoveBlockUpCommand = new RelayCommand(OnMoveBlockUpCommand);
-            MoveBlockDownCommand = new RelayCommand(OnMoveBlockDownCommand);
-            MoveBlockToTopCommand = new RelayCommand(OnMoveBlockToTopCommand);
-            MoveBlockToBottomCommand = new RelayCommand(OnMoveBlockToBottomCommand);
-        }
-        
-        public RelayCommand CopyBlockCommand { get; }
-        public RelayCommand PasteBlockCommand { get; }
-        public RelayCommand AddBlockCommand { get; }
-        public RelayCommand AddSectionCommand { get; }
-        public RelayCommand DeleteBlockCommand { get; }
-        public RelayCommand MoveBlockUpCommand { get; }
-        public RelayCommand MoveBlockDownCommand { get; }
-        public RelayCommand MoveBlockToTopCommand { get; }
-        public RelayCommand MoveBlockToBottomCommand { get; }
-
         public override void Initialise(IItemFilterBlockBase itemfilterBlock, IItemFilterScriptViewModel itemFilterScriptViewModel)
         {
             _parentScriptViewModel = itemFilterScriptViewModel;
@@ -56,52 +30,8 @@ namespace Filtration.ViewModels
             set
             {
                 ItemFilterCommentBlock.Comment = value;
+                RaisePropertyChanged();
             }
-        }
-
-        private void OnCopyBlockCommand()
-        {
-            _parentScriptViewModel.CopyBlock(this);
-        }
-
-        private void OnPasteBlockCommand()
-        {
-            _parentScriptViewModel.PasteBlock(this);
-        }
-
-        private void OnAddBlockCommand()
-        {
-            _parentScriptViewModel.AddBlock(this);
-        }
-
-        private void OnAddSectionCommand()
-        {
-            _parentScriptViewModel.AddCommentBlock(this);
-        }
-
-        private void OnDeleteBlockCommand()
-        {
-            _parentScriptViewModel.DeleteBlock(this);
-        }
-
-        private void OnMoveBlockUpCommand()
-        {
-            _parentScriptViewModel.MoveBlockUp(this);
-        }
-
-        private void OnMoveBlockDownCommand()
-        {
-            _parentScriptViewModel.MoveBlockDown(this);
-        }
-
-        private void OnMoveBlockToTopCommand()
-        {
-            _parentScriptViewModel.MoveBlockToTop(this);
-        }
-
-        private void OnMoveBlockToBottomCommand()
-        {
-            _parentScriptViewModel.MoveBlockToBottom(this);
         }
     }
 }
